refactor:  import path and send pageview with a custom path
