+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import styles from "./page.module.css";
 import About from "@/components/about/About";
 import Footer from "@/components/footer/Footer";
-import GoogleTracking from "@/components/goggleTracking/GoogleTracking";
+import GoogleTracking from "@/components/googleTracking/GoogleTracking";
 import Landing from "@/components/landing/Landing";
 import NavBar from "@/components/navBar/Navbar";
 import Projects from "@/components/projects/Projects";
