+++ src/components/googleTracking/GoogleTracking.tsx
@@ -5,6 +5,11 @@ import ReactGA from "react-ga4";
 const GoogleTracking = () => {
   useEffect(() => {
     ReactGA.initialize("G-FKQLHFB3VJ");
+    ReactGA.send({
+      hitType: "pageview",
+      page: "/",
+      title: "Selim Ben Hlel Portfolio",
+    });
   }, []);
   return null;
 };
